moved birdsnavigator to module
